Added 120px enemy sprite files to the sprite content folder with "shoot" and "death" animations ready to be hooked up.
Made slight improvements to enemy sprite animation and added 120px version to map file for previewing.

If it presents no techincal hurdles, I think we should use a single file name for the grouped triplets of sprite files (.aa, .json, .png) like I've done for the enemy sprite. It's easier to just make sure everything related has the same name, it's already pretty clear what type of file it is from the extension, and texturepacker already outputs the sprite sheet and the .json with the same name, so it's one less thing to manually fix for every new asset.